Rename enum label names for PG_Locale_Strategy
When PG_REGEX_BUILTIN was added it did not follow the same pattern as
the previous labels, i.e. PG_LOCALE_*, but in addition to this the two
libc strategies did not include in the name that they were realted to
libc. To fix both these issues we pick a new naming scheme:
PG_STRATEGY_<type>[_<subtype>] to make the code clearer.