Add LLVM_CPPFLAGS in headerscheck to llvm jit files
headerscheck on the CI currently works because the required llvm include
files are located in /usr/include (like /usr/include/llvm-c/Core.h).
However, it is very likely that those files will be located in a version
specific directory and llvm-config should be used to get the correct
cppflag using "$LLVM_CONFIG --cppflags".

Those flags are available in configure as LLVM_CPPFLAGS so we can add
them as extra flags for the headerscheck test command. This will allow
to run headerscheck on systems that don't have llvm include files in
/usr/include