Replace Latches with Interrupts
The Latch was a flag, with an inter-process signalling mechanism that
allowed a process to wait for the Latch to be set. My original vision
for Latches was that you could have different latches for different
things that you need to wait for, but but in practice, almost all code
used one "process latch" that was shared for all wakeups. The only
exception was the "recoveryWakeupLatch". I think the reason that we
ended up just sharing the same latch for all wakeups is that it was
cumbersome in practice to deal with multiple latches. For starters,
there was no machinery for waiting for multiple latches at the same
time. Secondly, changing the "ownership" of a latch needed extra
locking or other coordination. Thirdly, each inter-process latch
needed to be initialized at postmaster startup time.

This patch embraces the reality of how Latches were used, and replaces
the Latches with per-process interrupt mask. You can no longer
allocate Latches in arbitrary shared memory structs, an interrupt is
always directed at a particular process, addressed by its ProcNumber.
Each process has a bitmask of pending interrupts in PGPROC.

This commit introduces two interrupt bits. INTERRUPT_GENERAL_WAKEUP
replaces the general purpose per-process latch. All code that
previously set a process's process latch now sets its
INTERRUPT_GENERAL_WAKEUP interrupt bit instead.

The other interrupt bit, INTERRUPT_RECOVERY_WAKEUP, replaces
recoveryWakeupLatch. With this new interface, the code can easily wait
for the regular interrupts (INTERRUPT_GENERAL_WAKEUP) at the same time
as INTERRUPT_RECOVERY_WAKEUP. Previously, the code that waited on
recoveryWakeupLatch had to resort to timeouts to react to other
signals, because it was not possible to wait for two latches at the
same time. Previous attempt at unifying those wakeups was in commit
ac22929a26, but that was reverted in commit 00f690a239 because it
caused a lot of spurious wakeups when startup process was waiting for
recovery conflicts. The new machinery avoids that problem, by making
it easy to wait for two interrupts at the same time.

More interrupt bits are planned for followup patches, to replace the
various ProcSignal bits, as well as ConfigReloadPending and
ShutdownRequestPending.

This patch leaves behind a compatibility "latch.h" header, which
defines macros to map existing Latch functions to the corresponding
Interrupt functions. This reduces the code churn, and shows how we
could keep limited backwards-compatibility for extensions if
necessary.

This also moves the WaitEventSet functions to a different source file,
waiteventset.c. This separates the platform-dependent code waiting and
signalling code from the platform-independent parts.