Fix orphaned 2pc file which may casue instance restart failed
There is a race condition for global transaction when removing
2pc files:
   1. gxact1 is doing `FinishPreparedTransaction` and checkpoint
      is issued, so gxact1 will generate a 2pc file.
   2. then gxact1 was removed from `TwoPhaseState->prepXacts` and
      was put into freelist.
   3. but just before gxact1 remove its 2pc file, gxact2 is issued,
      gxact2 will reuse the same state memory of gxact1 and will
      reset `gxact->ondisk` to false.
   4. gxact1 continue and found that `gxact->ondisk` is false, it won't
      remove its 2pc file. This file is orphaned.

If gxact1's local xid is not frozen, the startup process will remove
the orphaned 2pc file. However, if the clog file is truncated by `vacuum`,
the startup process will raise error 'could not access status of transaction xxx',
due to it could not found the transaction's status file in dir `pg_xact`.