Runtime value of PG_TEST_EXTRA in Meson
Tests run using Meson ignore the value of environment variable
PG_TEST_EXTRA at the time of running tests and use the configuration
time value. In order to run the tests not specified in configuration
time PG_TEST_EXTRA, one needs to reconfigure which is costly. Allow
runtime value of PG_TEST_EXTRA to override the configuration time value.
In order to do this the configuration time value is passed as argument
"--pg-test-extra" to testwrap instead of adding it to the test
environment. If the environment variable is set at the time of running
test, its value is used. Otherwise value passed as the argument is used.

Meson builds use the same value of PG_TEST_EXTRA at the configuration
time and at the time of running the tests. Hence remove it from meson
setup command so as to treat both Meson and Make builds the same way in
CI.