Add PG_TEST_EXTRA configure option to the Make builds
The Meson builds have PG_TEST_EXTRA as a configure-time variable, which
was not available in the Make builds. To ensure both build systems are
in sync, PG_TEST_EXTRA is now added as a configure-time variable. It can
be set like this:
./configure PG_TEST_EXTRA="kerberos, ssl, ..."

Note that to preserve the old behavior, this configure-time variable is
overridden by the PG_TEST_EXTRA environment variable.

Author: Jacob Champion <jacob.champion@enterprisedb.com>
Reviewed by: Ashutosh Bapat, Nazir Bilal Yavuz with inputs from Tom Lane
and Andrew Dunstan