Improve comment about configuration time PG_TEST_EXTRA
The comment mentioned "not safe" and "while-space separate list" which are not
relevant to the code in meson.build file. The purpose mentioned in the comment
has already gone stale; the variable is used for tests which take longer but are
otherwise safe. And the content may change in the future. This code should just
make the value of the variable available to the tests irrespective of its
purpose and content. Improved the comment to mention how the PG_TEST_EXTRA is
passed to tests instead of the variable's content and purpose.

Ashutosh Bapat