Use RestrictInfo hash table for storing EC derived child RestrictInfos
RestrictInfos derived from an EquivalenceClass, including the child
RestrictInfos, are stored in its EquivalenceClass::ec_derives which
is a linked list. The linked list is scanned every time
create_join_clause() is called, which in turn is called multiple times
for a given partition. The loop thus contributes O(N^2) at run time
where N is number of partitions. When there are thousands of
partition, the loops contributes significantly to the planning time.

In order to reduce the time complexity of searching a clause, store the
child clauses in a hash table, so that contribution of this loop in
overall planning time reduces to O(N). Please note that the parent
derived clauses are still stored in EquivalenceClass::ec_derives.

The hash table approach differs from the linked list based approach in
the following aspects:

1. The linked lists are per EquivalenceClass whereas we use a single
   hash table for all the child clauses. A hash table consumes more
   memory compared to the linked list, thus having one hash table per
   EquivalenceClass causes more memory to be consumed. But hash tables
   can efficiently handle thousands of entries hence one hash table to
   store all the child derived clauses from all the EquivalenceClasses
   suffices.

2. EquivalenceClass::ec_derives is searched by
   EquivalenceClass::left_em/right_em pointers whereas the hash table is
   key'ed by RestrictInfo::rinfo_serial and
   RestrictInfo::required_relids. In the future (next commits) the hash
   table will be used for storing the child RestrictInfos produced by
   translating the corresponding parent RestrictInfos which may not be
   part of EquivalenceClasses. Such RestrictInfos do not have
   EquivalenceMembers associated with them and hence can not be used as
   a key. OTOH, RestrictInfo::rinfo_serial and
   RestrictInfo::required_relids is set in all RestrictInfos.

   To cope with this, create_join_clause() searches for the parent
   clause in EquivalenceClass::ec_sources or
   EquivalenceClass::ec_derives before searching for the child clause in
   the hash table. That's some extra work compared to linked list based
   approach where a child clause could be searched directly in
   EquivalenceClass::ec_derives. But with this change, the linked lists
   should be short and thus searching for a parent clause should not
   take much time.

Please note that EquivalenceClass:ec_derives is used in functions other
than create_join_clause(). But all those usages happen before any child
EquivalenceMembers are added. This commit adds Asserts in those places
or leverages existing Asserts to validate this fact. Also the Asserts
will help in noticing any change in case the assumption is broken in the
future.  Hence moving child derived clauses out of EquivalenceClass does
not need any changes to those functions.

Ashutosh Bapat