Add XMLSerialize: version and explicit XML declaration
* XMLSerialize: explicit XML declaration (SQL/XML X078)

This adds the options INCLUDING XMLDECLARATION and EXCLUDING
XMLDECLARATION to XMLSerialize, which includes or remove the
XML declaration of the given DOCUMENT or CONTENT, respectively.
If not specified, the output will contain an XML declaration
only if the given XML value had one.

* XMLSerialize: version (SQL/XML X076)

The option VERSION can be used to specify the version in the
XML declaration of the serialized DOCUMENT or CONTENT. In case
of XML values of type DOCUMENT, the version will be validated
by libxml2's xmlNewDoc(), which will raise an error for invalid
versions or a warning for unsupported ones. For CONTENT values
no validation is performed. This option is ignored if the XML
value had no XML deldaration and INCLUDING XMLDECLARATION was
not used.

Usage examples:

SELECT
  xmlserialize(
    DOCUMENT xmlval AS text
    VERSION '1.0'
    INCLUDING XMLDECLARATION);

SELECT
  xmlserialize(
    DOCUMENT xmlval AS text
    EXCLUDING XMLDECLARATION);

This patch also includes regression tests and documentation.