Introduce VALUES -> ARRAY transformation.
Transform accidentally appearing 'x IN (VALUES, ...) expressions to
x IN 'ANY ...'. The second variant is better because it lets the planner avoid
one more unnecessary SEMI JOIN operator.

This form of expression usually appears in auto-generated queries as a corner
case of searching an object in a set of other ones when the object is described
by only one property.

Let this unusual optimisation be in the core because the planner would only
spend a few more cycles without this construct.