Move check for binary mode and on_error option to appropriate location
The current location for checking for binary mode and on_error option
is not appropriate, as this is where checks should be performed before
inserting default values.
Additionally, the comment mentions two checks, but there are now three.

This patch moves the check for incompatible options to the end of
ProcessCopyOptions() and updates the comment accordingly.