Compact PlannerInfo to restore its previous size
Previous commit added a new member to PlannerInfo increasing its size
from 696 to 704 bytes in size on my laptop. This change coincides with a
potential planning time regression for query involving partitioned
tables with lower number of partitios (around 100 or less) and lower
number of joins between partitioned tables. This commit rearranges the
members of PlannerInfo so that the size is back to 696. We should merge
this commit into previous commit in case this fixes the regression,
otherwise discard it.

Ashutosh Bapat