Introduce "ALL SEQUENCES" support for PostgreSQL logical replication
This commit enhances logical replication by enabling the inclusion of all
sequences in publications.

Furthermore, enhancements to psql commands (\d and \dRp) now allow for better
display of publications containing specific sequences or sequences included
in a publication.

Note: This patch currently supports only the "ALL SEQUENCES" clause.
Handling of clauses such as "FOR SEQUENCE" and "FOR SEQUENCES IN SCHEMA"
will be addressed in a subsequent patch.