Enable support for 'publish_generated_columns' option.
Generated column values are not currently replicated because it is assumed
that the corresponding subscriber-side table will generate its own values
for those columns.

This patch supports the transmission of generated column information and data
alongside regular table changes. This behaviour is partly controlled by a new
publication parameter 'publish_generated_columns'.

Example usage:
CREATE PUBLICATION pub1 FOR TABLE tab_gencol WITH (publish_generated_columns = true);

Generated columns can also be published if they are specified in a
publication column list. This overrides the parameter, so it works even if
'publish_generated_columns' is false.

When the subscription parameter 'copy_data' is true, then data is copied
during the initial table synchronization using the COPY command. The
normal COPY command does not copy generated columns, so if generated columns are
published we need to use a different form of the copy syntax:
'COPY (SELECT column_name FROM table_name) TO STDOUT'.

~~

Behavior Summary:

A. when generated columns are published

* Publisher not-generated column => subscriber not-generated column:
  This is just normal logical replication (not changed by this patch).

* Publisher not-generated column => subscriber generated column:
  This will give ERROR.

* Publisher generated column => subscriber not-generated column:
  The publisher generated column value is copied.

* Publisher generated column => subscriber generated column:
  This will give ERROR.

B. when generated columns are not published

* Publisher not-generated column => subscriber not-generated column:
  This is just normal logical replication (not changed by this patch).

* Publisher not-generated column => subscriber generated column:
  This will give ERROR.

* Publisher generated column => subscriber not-generated column:
  The publisher generated column is not replicated. The subscriber column
  will be filled with the subscriber-side default data.

* Publisher generated column => subscriber generated column:
  The publisher generated column is not replicated. The subscriber
  generated column will be filled with the subscriber-side computed or
  default data.

~

There is a change in 'pg_publication' catalog so we need to
bump the catversion.