Add support for NOT ENFORCED in foreign key constraints.
Normally, when a foreign key (FK) constraint is created on a table,
action/check triggers are associated with the constraint to ensure data
integrity. With this patch, when constraints are marked as NOT
ENFORCED, integrity checks are not required, and thus, these triggers
are unnecessary. As a result, when creating a NOT ENFORCED FK
constraint or changing an existing FK constraint to NOT ENFORCED, the
creation of triggers will be skipped, or existing triggers will be
dropped, respectively. And when changing an existing NOT ENFORCED FK
constraint to ENFORCED, those triggers will be created.