Add regression tests.
As this patch series adds the CONCURRENTLY option to the VACUUM FULL and
CLUSTER commands, it's appropriate to test that the "concurrent data changes"
(i.e. changes done by application while we are copying the table contents to
the new storage) are processed correctly.

Injection points are used to stop the data copying at some point. While the
backend in charge of the copying is waiting on the injection point, another
backend runs some INSERT, UPDATE and DELETE commands on the table. Then we
wake up the first backend and let the CLUSTER (CONCURRENTLY) command
finish. Finally we check that all the "concurrent data changes" are present in
the table and that they contain the correct visibility information.