Lift and Shift Bitmap Table Scan prefetch functions
This is a temporary commit. For ease of review, move the Bitmap Table
Scan prefetch functions into heapam_handler.c without modifying them.
The next commit will push the prefetching code into heap-specific code.
Separating the commit which moves to the functions to heapam_handler.c
and the one which modifies them and their callers to contain bitmap
prefetch code in heap AM-specific code makes it easier to see what has
changed. Doing so requires including heap code in nodeBitmapHeapscan.c,
which is not okay for the final code that is committed.