Implements last_update_wins conflict resolver.
This resolver is applicable for conflict types: insert_exists, update_exists,
update_origin_differs and delete_origin_differs.

For these conflicts, when the resolver is set to last_update_wins,
the timestamps of the remote and local conflicting tuple are compared to
determine whether to apply or ignore the remote changes.
The GUC track_commit_timestamp must be enabled to support this resolver.

Since conflict resolution for two phase commit transactions using
prepare-timestamp can result in data divergence, this patch restricts
enabling both two_phase and the last_update_wins resolver together
for a subscription.

The patch also restrict starting a parallel apply worker if resolver is set
to last_update_wins for any conflict type.