Avoid full btree index scans when skipping is possible
Previously, we could ignore the skip signal until the end of the range of
values producable by the index scan key. Now, we can fail to start a new
primscan only for up to number of parallel workers + 1 buffers, at the cost
of doing a bit more before releasing the scan while we process the 'we may
need a new primitive scan' signal.

If we detect that a parallel worker in the same primscan range thinks
this is the right moment to start a new primitive scan, we don't release
the parallel scan immediately, but instead only release it after reading
the pages contents to find out if we really should start a new primitive
scan.  If so, we start that new primitive scan, and if instead we find
we've already skidded into the range of pages we would've arrived on with
the skip scan, we instead mark that the primitive scan has reached a new
primscan range, do some cleanup, and then continue the scan as usual.